+++ packages/grid/src/features/standard-table/components/StandardTableContent.tsx
@@ -1,6 +1,7 @@
-import * as React from "react";
-import { Row, Spacing, Text } from "@stenajs-webui/core";
+import { Row, Spacing } from "@stenajs-webui/core";
+import { Banner } from "@stenajs-webui/elements";
 import { ErrorScreen, LoadingScreen } from "@stenajs-webui/panels";
+import * as React from "react";
 import { StandardTableProps, StandardTableVariant } from "./StandardTable";
 import { StandardTableRowList } from "./StandardTableRowList";
 
@@ -25,6 +26,9 @@ export const StandardTableContent = React.memo(function StandardTableContent<
   loading,
   items,
   noItemsLabel = "There is no data available.",
+  noItemsContentRight,
+  noItemsContentBottom,
+  noItemsHeader,
   colIndexOffset,
   rowIndexOffset,
   variant,
@@ -47,8 +51,15 @@ export const StandardTableContent = React.memo(function StandardTableContent<
 
   if (!items || !items.length) {
     return (
-      <Row spacing={10} justifyContent={"center"}>
-        <Text>{noItemsLabel}</Text>
+      <Row spacing={2} justifyContent={"center"}>
+        <Banner
+          text={noItemsLabel}
+          headerText={noItemsHeader}
+          contentRight={noItemsContentRight}
+          variant={"info"}
+        >
+          {noItemsContentBottom}
+        </Banner>
       </Row>
     );
   }